:recyle: app restructing
